Tile values are casted to int while unmerging
In order to avoid odd float representation of the tile values, the
result of the diversion is casted to int while undoing the merge
command.

+++ model/commands/board.py
@@ -63,7 +63,7 @@ class MergeTile(BoardTile):
         self.board.remove(self.source)
 
     def undo(self):
-        self.destinationTile.value = self.destinationTile.value / 2
+        self.destinationTile.value = int(self.destinationTile.value / 2)
         self.board.add(self.source, self.sourceTile)
